Custom Authority Extraction
What authorites are actually implied in a JWT can vary by OAuth
provider. The attribute name and the authority representation can
differ. While perhaps the most typical behavior is a "scope"
attribute, it isn't always this way.

Authority extraction also allows the user to derive authorities and
roles within the domain of the application, for example by removing
the SCOPE prefix in order to remain OAuth-agnostic or by querying
roles from a database.