Unsecured Resource Server Artifact
The demo will be more efficient if we aren't changing the project
dependencies in the IDE. However, it will be more compelling in the
beginning if the app completely lacks any security at all, meaning
we need to leave the dependencies out.

So, this article has all the dependencies missing and is what we
run at the start of the presentation. Then, when we switch over to
the IDE, we stop this application and start the one in the IDE that
has the dependencies there waiting for us.